FIX: Problems with accuracy of progress bar

+++ aTunes/src/main/java/net/sourceforge/atunes/kernel/modules/player/mplayer/MPlayerEngine.java
@@ -251,7 +251,7 @@ public class MPlayerEngine extends AbstractPlayerEngine {
     }
 
     @Override
-    protected void seekTo(double position) {
+    protected void seekTo(long position) {
         commandWriter.sendSeekCommand(position);
     }
